Move SingleOutput to shared
